Change the authentication logic on LDAP
Can't just use a filter to get users that are member of the target group,
since the users don't have a "memberOf" attribute set on them.
Had to implement a hierarchical search of groups that belongs to the target
group and check whether at least one of these groups contains the user that
is trying to sign in.